Rename checker.h to control_data_checker.h
